move IconInDivCganged to helpers module
